requirements.txt
openai
tiktoken
psycopg2
google-generativeai